Add Watson distribution simulation code
Separate functions for simulating Watson bipolar and girdle data.  Also
a function to simulate Watson data which calls the appropriate function
based on the value of the argument kappa.

An internal function addpoint.Watson has been added to support these
functions.  This reason for this is that the algorithms for simulating
Watson bipolar and girdle data both simulate values, check whether those
values satisfy criteria and if they do, determine a Watson variate based
on those values.   The role of addpoint.Watson is to determine the
Watson variates based on simulated values that satisfy the criteria.