Marker indicating in.W.poly (currently not used by in.W) still needs some investigation and is not to be used at this stage.
Needs tweaking to handle cases where a point being checked is within W
but the distance within the polygon to at least one vertex exceeds
pi*rad.  This is a low priority fix given that the Bevis and Chatelain
algorithm works and is currently used by the function in.W that users
should use.